+++ examples/resources/sendgrid_bounce_settings/resource.tf
@@ -1,0 +1,3 @@
+resource "sendgrid_bounce_settings" "example" {
+  soft_bounce_purge_days = 30
+}